bot/app/pyproject.toml
[tool.poetry]
name = "rlt-test"
version = "0.0.1"
description = "rlt-test"
authors = ["Konstantin Klepikov <oformleno@gmail.com>"]

[tool.poetry.dependencies]
python = "3.10.6"
passlib = "1.7.4"
toml = "0.10.2"
aiogram = "3.1.1"
pydantic = "^2.3"
aiogram-dialog = "2.0.0"
pydantic-settings = "^2.0.3"
python-dotenv = "1.0.0"
motor = "^3.1.1"
watchdog = {extras = ["watchmedo"], version = "^3.0.0"}
python-dateutil = "^2.8.2"

[tool.poetry.group.dev.dependencies]
pytest = ">=4.6"
ipython = "^8.0.0"
mypy = "^1.0.1"
black = "^21.12b0"
isort = "^5.10.1"
autoflake = "^1.4"
flake8 = "^4.0.1"
pytest-cov = "^3.0.0"
pytest-asyncio = "^0.20.3"
click = "^8.1.3"
watchdog = "^3.0.0"

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = [
    "tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]
asyncio_mode = "auto"
